editor/libeditor/html/tests/test_bug601881.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=601881
-->
<head>
  <title>Test for Bug 601881</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=601881">Mozilla Bug 601881</a>
<p id="display"></p>
<div id="content">
  <video id="v"></video>
  <audio id="a"></audio>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 601881 **/

var v = document.getElementById("v");
var a = document.getElementById("a");
var content = document.getElementById("content");

function runPass(makeEditable, makeUneditable, callback) {
  is(v.mozAutoplayEnabled, true, "Video should be autoplay-able when not made editable yet");
  is(a.mozAutoplayEnabled, true, "Audio should be autoplay-able when not made editable yet");
  makeEditable();
  is(v.mozAutoplayEnabled, false, "Video should not be autoplay-able when made editable");
  is(a.mozAutoplayEnabled, false, "Audio should not be autoplay-able ehen made editable");
  makeUneditable();
  is(v.mozAutoplayEnabled, true, "Video should be autoplay-able after being made uneditable");
  is(a.mozAutoplayEnabled, true, "Audio should be autoplay-able after being made uneditable");

  callback();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  runPass(function () document.designMode = 'on',
          function () document.designMode = 'off',
          function () {
            runPass(function() content.setAttribute("contenteditable", "true"),
                    function() content.removeAttribute("contenteditable"),
                    function() SimpleTest.finish()
            )
          }
  );
});

</script>
</pre>
</body>
</html>
